+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [2.0.0] - 2026-06-02
+
 ### Changed
 
 - Renamed the `version` subcommand to `get` to avoid confusion with the `--version` flag. Use `gitsemver get`
@@ -166,7 +168,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Functions signature for `EnsureUpToDate` and `ResolveVersion`.
 
-[Unreleased]: https://github.com/giantswarm/gitsemver/compare/v1.1.2...HEAD
+[Unreleased]: https://github.com/giantswarm/gitsemver/compare/v2.0.0...HEAD
+[2.0.0]: https://github.com/giantswarm/gitsemver/compare/v1.1.2...v2.0.0
 [1.1.2]: https://github.com/giantswarm/gitsemver/compare/v1.1.1...v1.1.2
 [1.1.1]: https://github.com/giantswarm/gitsemver/compare/v1.1.0...v1.1.1
 [1.1.0]: https://github.com/giantswarm/gitsemver/compare/v1.0.2...v1.1.0